Fix vendor:publish "Can't locate path: <0>" bug
Was getting a bug, 'Can't locate path: <0>', when running 'php artisan vendor:publish --provider="Codecasts\Auth\JWT\ServiceProvider"'. This is the fix to that error. Has not been tested on multiple versions of Laravel but looks simple enough. The publishes method should take a key => value array, not a simple text array as was previously in the code.